Maven publishing configuration setup
This change has removed older jaring scripts, and some distribution
libraries in favor of Maven creating a fully working deployed jar. Note
that Android Studio restricts duplicate classes from being in a project
at all, which forces java3d-utils-and to merge it's target jar with
java3-util as well as suppress publication of the non "-android" version
of the joal libraries.

Signed-off-by: Philip Jordan <p.j.nz@outlook.com>